wraps dashboard in protected route, aadds dashboard layout
