+++ src/assets/icons/Icons.jsx
@@ -210,3 +210,62 @@ export function Key(props) {
     </svg>
   );
 }
+export function SettingsGear(props) {
+  return (
+    <svg
+      width="46"
+      height="46"
+      viewBox="0 0 46 46"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+      {...props}
+    >
+      <path d="M22.0983 16.6864C20.8496 16.6864 19.6289 17.0566 18.5906 17.7504C17.5523 18.4442 16.7431 19.4302 16.2652 20.5839C15.7873 21.7376 15.6623 23.0071 15.9059 24.2318C16.1495 25.4565 16.7509 26.5815 17.6339 27.4645C18.5168 28.3475 19.6418 28.9488 20.8666 29.1924C22.0913 29.4361 23.3608 29.311 24.5145 28.8331C25.6681 28.3553 26.6542 27.546 27.348 26.5078C28.0417 25.4695 28.412 24.2488 28.412 23.0001C28.4098 21.3262 27.7439 19.7216 26.5604 18.538C25.3768 17.3544 23.7721 16.6885 22.0983 16.6864ZM22.0983 27.6667C21.1753 27.6667 20.2731 27.393 19.5057 26.8802C18.7382 26.3675 18.1401 25.6386 17.7869 24.7859C17.4337 23.9332 17.3413 22.9949 17.5213 22.0896C17.7014 21.1844 18.1459 20.3529 18.7985 19.7002C19.4511 19.0476 20.2827 18.6031 21.1879 18.4231C22.0931 18.243 23.0314 18.3354 23.8842 18.6886C24.7369 19.0418 25.4657 19.64 25.9785 20.4074C26.4913 21.1748 26.765 22.0771 26.765 23.0001C26.765 24.2377 26.2733 25.4247 25.3981 26.2999C24.523 27.175 23.336 27.6667 22.0983 27.6667ZM36.9218 20.2001C36.8979 20.0853 36.8498 19.977 36.7807 19.8823C36.7116 19.7876 36.623 19.7088 36.521 19.651L32.343 17.2656L32.2853 17.1681L32.2689 12.4589C32.2689 12.34 32.2431 12.2224 32.1933 12.1144C32.1436 12.0063 32.071 11.9104 31.9806 11.833C30.5214 10.5992 28.8415 9.65331 27.0299 9.04541C26.9209 9.00879 26.8056 8.99537 26.6912 9.00601C26.5768 9.01664 26.4658 9.0511 26.3655 9.10717L22.1615 11.4515H22.0242L17.8311 9.10443C17.7303 9.04819 17.6189 9.01376 17.504 9.00335C17.3891 8.99295 17.2732 9.00681 17.164 9.04404C15.3526 9.65371 13.6734 10.602 12.216 11.8385C12.1253 11.9154 12.0523 12.0111 12.0021 12.1189C11.9518 12.2267 11.9256 12.3441 11.925 12.463L11.9031 17.1764C11.8838 17.2079 11.8646 17.2409 11.8468 17.2738L7.67426 19.6469C7.57172 19.7055 7.48298 19.7855 7.41407 19.8814C7.34515 19.9773 7.29767 20.0869 7.27485 20.2028C6.90838 22.0506 6.90838 23.9523 7.27485 25.8C7.29873 25.9148 7.34684 26.0232 7.41595 26.1178C7.48507 26.2125 7.5736 26.2913 7.67563 26.3491L11.8509 28.7277L11.9086 28.8251L11.925 33.5343C11.9244 33.6546 11.9501 33.7736 12.0003 33.883C12.0506 33.9923 12.1242 34.0893 12.216 34.1671C13.6752 35.4009 15.3551 36.3468 17.1668 36.9547C17.2757 36.9913 17.391 37.0047 17.5055 36.9941C17.6199 36.9835 17.7308 36.949 17.8311 36.8929L22.0324 34.5486H22.1697L26.3642 36.8957C26.4869 36.9635 26.6247 36.9994 26.765 37C26.8551 37.0004 26.9447 36.9855 27.0299 36.9561C28.8389 36.3456 30.5156 35.3973 31.971 34.1616C32.0617 34.0847 32.1347 33.989 32.185 33.8812C32.2352 33.7734 32.2615 33.656 32.262 33.5371L32.284 28.8238C32.3032 28.7922 32.3224 28.7592 32.3402 28.7263L36.5155 26.3532C36.6183 26.2947 36.7073 26.2148 36.7765 26.1189C36.8456 26.023 36.8933 25.9133 36.9163 25.7973C37.2846 23.9499 37.2864 22.0482 36.9218 20.2001ZM35.3777 25.1124L31.3383 27.4059C31.2073 27.4809 31.0991 27.59 31.0253 27.7216C30.9444 27.8589 30.8593 28.0112 30.77 28.1526C30.6888 28.2827 30.6456 28.4329 30.6451 28.5863L30.6232 33.1445C29.4875 34.0544 28.2119 34.7742 26.8459 35.2761L22.7846 33.0031C22.6618 32.9341 22.5233 32.8977 22.3824 32.8975H22.3632C22.193 32.8975 22.0201 32.8975 21.8499 32.8975C21.7026 32.8939 21.5571 32.9299 21.4285 33.0018L17.3562 35.2747C15.9885 34.7754 14.711 34.0574 13.5735 33.1486L13.557 28.5959C13.5565 28.4425 13.5133 28.2923 13.4321 28.1622C13.3442 28.0249 13.2578 27.8767 13.1768 27.7312C13.1038 27.5995 12.996 27.4903 12.8652 27.4155L8.82582 25.1138C8.59101 23.7149 8.59101 22.2866 8.82582 20.8877L12.8584 18.5942C12.9914 18.52 13.1015 18.4109 13.1768 18.2785C13.2578 18.1413 13.3429 17.9889 13.4321 17.8475C13.5133 17.7174 13.5565 17.5672 13.557 17.4138L13.5789 12.8556C14.713 11.9462 15.9867 11.2264 17.3507 10.724L21.412 12.997C21.5402 13.07 21.686 13.1061 21.8334 13.1013C22.0036 13.1013 22.1765 13.1013 22.3467 13.1013C22.494 13.1048 22.6395 13.0688 22.7681 12.997L26.8377 10.724C28.2054 11.2233 29.4829 11.9414 30.6204 12.8501L30.6369 17.4028C30.6373 17.5562 30.6806 17.7064 30.7618 17.8365C30.8496 17.9738 30.9361 18.122 31.0171 18.2675C31.0901 18.3993 31.1979 18.5085 31.3287 18.5832L35.3708 20.8863C35.607 22.285 35.6084 23.7133 35.3749 25.1124H35.3777Z" />
+    </svg>
+  );
+}
+
+export function ChartBar(props) {
+  return (
+    <svg
+      width="46"
+      height="46"
+      viewBox="0 0 46 46"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+      {...props}
+    >
+      <path d="M34.844 32.6436H33.594V12.3936C33.594 12.1946 33.515 12.0039 33.3743 11.8632C33.2337 11.7226 33.0429 11.6436 32.844 11.6436H25.844C25.6451 11.6436 25.4543 11.7226 25.3137 11.8632C25.173 12.0039 25.094 12.1946 25.094 12.3936V17.6436H18.844C18.6451 17.6436 18.4543 17.7226 18.3137 17.8632C18.173 18.0039 18.094 18.1946 18.094 18.3936V23.6436H12.844C12.6451 23.6436 12.4543 23.7226 12.3137 23.8632C12.173 24.0039 12.094 24.1946 12.094 24.3936V32.6436H10.844C10.6451 32.6436 10.4543 32.7226 10.3137 32.8632C10.173 33.0039 10.094 33.1946 10.094 33.3936C10.094 33.5925 10.173 33.7832 10.3137 33.9239C10.4543 34.0645 10.6451 34.1436 10.844 34.1436H34.844C35.0429 34.1436 35.2337 34.0645 35.3743 33.9239C35.515 33.7832 35.594 33.5925 35.594 33.3936C35.594 33.1946 35.515 33.0039 35.3743 32.8632C35.2337 32.7226 35.0429 32.6436 34.844 32.6436ZM26.594 13.1436H32.094V32.6436H26.594V13.1436ZM19.594 19.1436H25.094V32.6436H19.594V19.1436ZM13.594 25.1436H18.094V32.6436H13.594V25.1436Z" />
+    </svg>
+  );
+}
+
+export function Profile(props) {
+  return (
+    <svg
+      width="46"
+      height="46"
+      viewBox="0 0 46 46"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+      {...props}
+    >
+      <path d="M35.3655 33.8725C33.432 30.5313 30.3965 28.1868 26.8694 27.1925C28.5842 26.3208 29.9554 24.897 30.762 23.1506C31.5687 21.4043 31.7636 19.4371 31.3154 17.5664C30.8672 15.6957 29.8021 14.0304 28.2917 12.8392C26.7813 11.6479 24.9137 11 22.99 11C21.0664 11 19.1988 11.6479 17.6884 12.8392C16.178 14.0304 15.1128 15.6957 14.6646 17.5664C14.2165 19.4371 14.4114 21.4043 15.218 23.1506C16.0246 24.897 17.3959 26.3208 19.1107 27.1925C15.5836 28.1856 12.5481 30.5301 10.6145 33.8725C10.5614 33.9561 10.5258 34.0495 10.5097 34.1471C10.4937 34.2448 10.4975 34.3447 10.5211 34.4408C10.5446 34.5369 10.5874 34.6273 10.6467 34.7065C10.7061 34.7856 10.7809 34.852 10.8666 34.9015C10.9523 34.951 11.0471 34.9827 11.1453 34.9946C11.2436 35.0066 11.3432 34.9985 11.4382 34.9709C11.5333 34.9434 11.6218 34.8969 11.6984 34.8342C11.775 34.7715 11.8381 34.694 11.884 34.6063C14.2334 30.5472 18.3842 28.1245 22.99 28.1245C27.5958 28.1245 31.7467 30.5472 34.096 34.6063C34.1419 34.694 34.2051 34.7715 34.2817 34.8342C34.3583 34.8969 34.4468 34.9434 34.5418 34.9709C34.6368 34.9985 34.7365 35.0066 34.8347 34.9946C34.933 34.9827 35.0278 34.951 35.1135 34.9015C35.1992 34.852 35.2739 34.7856 35.3333 34.7065C35.3927 34.6273 35.4354 34.5369 35.459 34.4408C35.4825 34.3447 35.4864 34.2448 35.4703 34.1471C35.4543 34.0495 35.4186 33.9561 35.3655 33.8725ZM15.8967 19.5635C15.8967 18.1606 16.3127 16.7891 17.0921 15.6226C17.8715 14.4561 18.9794 13.547 20.2755 13.0101C21.5717 12.4732 22.9979 12.3327 24.3739 12.6064C25.7499 12.8801 27.0138 13.5557 28.0058 14.5477C28.9978 15.5397 29.6734 16.8037 29.9471 18.1796C30.2208 19.5556 30.0803 20.9819 29.5434 22.278C29.0066 23.5742 28.0974 24.682 26.9309 25.4614C25.7644 26.2408 24.393 26.6569 22.99 26.6569C21.1094 26.6546 19.3065 25.9065 17.9768 24.5768C16.647 23.247 15.8989 21.4441 15.8967 19.5635Z" />
+    </svg>
+  );
+}
+
+export default function House(props) {
+  return (
+    <svg
+      width="45"
+      height="45"
+      viewBox="0 0 45 45"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+      {...props}
+    >
+      <path d="M24.155 8.78087L24.33 8.94487L37.402 21.7869L36 23.2129L34.2 21.4449L34.201 34.9999C34.201 36.0539 33.385 36.9179 32.35 36.9939L32.201 36.9999H12.201C11.147 36.9999 10.283 36.1839 10.206 35.1489L10.201 34.9999L10.2 21.4459L8.402 23.2129L7 21.7869L20.058 8.95787C21.171 7.82187 22.966 7.75887 24.155 8.78087ZM21.569 10.2849L21.473 10.3719L12.2 19.4809L12.201 34.9999L17.2 34.9989L17.201 24.9999C17.201 23.9459 18.017 23.0819 19.052 23.0049L19.201 22.9999H25.201C26.255 22.9999 27.119 23.8159 27.196 24.8509L27.201 24.9999L27.2 34.9989L32.201 34.9999L32.2 19.4799L22.901 10.3439C22.537 9.98687 21.969 9.96487 21.569 10.2849ZM25.201 24.9999H19.201L19.2 34.9989H25.2L25.201 24.9999Z" />
+    </svg>
+  );
+}
