+++ src/routes/Router.jsx
@@ -5,8 +5,9 @@ import SpesificVenueSupabase from "../pages/SpesificVenueSupabase";
 import SearchMap from "../pages/SearchMap";
 import Login from "../pages/Login";
 import Signup from "../pages/SignUp";
-import Dashboard from "../pages/Dashboard";
 import ProtectedRoute from "./ProtectedRoute";
+import DashboardIndex from "../pages/dashboard/DashboardIndex";
+import DashboardLayout from "../pages/dashboard/DashboardLayout";
 
 export default function Router() {
   return (
@@ -18,14 +19,12 @@ export default function Router() {
         <Route path="/login" element={<Login />} />
         <Route path="/signup" element={<Signup />} />
         <Route path="/venue/supabase/:id" element={<SpesificVenueSupabase />} />
-        <Route
-          path="/dashboard"
-          element={
-            <ProtectedRoute>
-              <Dashboard />
-            </ProtectedRoute>
-          }
-        />
+        <Route element={<ProtectedRoute />}>
+          <Route path="/dashboard" element={<DashboardLayout />}>
+            <Route index element={<DashboardIndex />} />
+            <Route path="/dashboard/profile" element={<DashboardIndex />} />
+          </Route>
+        </Route>
       </Routes>
     </>
   );
